# specs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API enables end-users, owners and Sally devices to interact  with the central database.",
    "version": "1.0",
    "title": "Pantry REST API",
    "termsOfService": "http://casabots.com/terms/",
    "contact": {},
    "license": {
      "name": "Copyright (c) CASABOTS 2016"
    }
  },
  "host": "localhost:8080",
  "basePath": "/pantry",
...
  "paths": {
...
    "/sally": {
      "post": {
        "tags": [
          "sally-controller"
        ],
        "summary": "createSally",
        "operationId": "createSallyUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "sally",
            "description": "sally",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Sally"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Sally"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/sally/by-ingredient": {
      "get": {
        "tags": [
          "sally-controller"
        ],
        "summary": "getSallyDevicesByIngredient",
        "operationId": "getSallyDevicesByIngredientUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ingredient",
            "in": "query",
            "description": "ingredient",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Sally"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/sally/by-location": {
      "get": {
        "tags": [
          "sally-controller"
        ],
        "summary": "getSallyDevicesByLocation",
        "operationId": "getSallyDevicesByLocationUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "description": "latitude",
            "required": true,
            "type": "number",
            "format": "double"
          },
          {
            "name": "longitude",
            "in": "query",
            "description": "longitude",
            "required": true,
            "type": "number",
            "format": "double"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Sally"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/sally/by-salad": {
      "get": {
        "tags": [
          "sally-controller"
        ],
        "summary": "getSallyDevicesBySalad",
        "operationId": "getSallyDevicesBySaladUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "salad",
            "in": "query",
            "description": "salad",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Sally"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/sally/generate-code": {
      "get": {
        "tags": [
          "sally-controller"
        ],
        "summary": "generateUniqueCode",
        "operationId": "generateUniqueCodeUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/sally/login": {
      "post": {
        "tags": [
          "sally-controller"
        ],
        "summary": "login",
        "operationId": "loginUsingPOST_2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "sally",
            "description": "sally",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/sally/pending-order": {
      "get": {
        "tags": [
          "sally-controller"
        ],
        "summary": "getPendingOrder",
        "operationId": "getPendingOrderUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/sally/update": {
      "post": {
        "tags": [
          "sally-controller"
        ],
        "summary": "updateSally",
        "operationId": "updateSallyUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "sally",
            "description": "sally",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Sally"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Sally"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/sally/{sallyId}": {
      "get": {
        "tags": [
          "sally-controller"
        ],
        "summary": "getSally",
        "operationId": "getSallyUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "sallyId",
            "in": "path",
            "description": "sallyId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Sally"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "sally-controller"
        ],
        "summary": "deleteSally",
        "operationId": "deleteSallyUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "sallyId",
            "in": "path",
            "description": "sallyId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/sally/{sallyId}/ingredients": {
      "get": {
        "tags": [
          "sally-controller"
        ],
        "summary": "getIngredients",
        "operationId": "getIngredientsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "sallyId",
            "in": "path",
            "description": "sallyId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Ingredient"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/sally/{sallyId}/salads": {
      "get": {
        "tags": [
          "sally-controller"
        ],
        "summary": "getSalads",
        "operationId": "getSaladsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "sallyId",
            "in": "path",
            "description": "sallyId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Salad"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/sizes": {
      "post": {
        "tags": [
          "size-controller"
        ],
        "summary": "createSize",
        "operationId": "createSizeUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "size",
            "description": "size",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Size"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/sizes/update": {
      "post": {
        "tags": [
          "size-controller"
        ],
        "summary": "updateSize",
        "operationId": "updateSizeUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "size",
            "description": "size",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Size"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/sizes/{sizeId}": {
      "get": {
        "tags": [
          "size-controller"
        ],
        "summary": "getSize",
        "operationId": "getSizeUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "sizeId",
            "in": "path",
            "description": "sizeId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Size"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "size-controller"
        ],
        "summary": "deleteSize",
        "operationId": "deleteSizeUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "sizeId",
            "in": "path",
            "description": "sizeId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Order": {
      "type": "object",
      "properties": {
        "customSalad": {
          "type": "boolean"
        },
        "orderId": {
          "type": "string"
        },
        "orderPlaced": {
          "type": "boolean"
        },
        "placedTime": {
          "type": "string",
          "format": "date-time"
        },
        "saladId": {
          "type": "string"
        },
        "sallyId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      }
    },
...
    "Size": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "sizeName": {
          "type": "string"
        }
      }
    },
...
    "LoginUser": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      }
    },
...
    "Sally": {
      "type": "object",
      "properties": {
        "ingredients": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "location": {
          "$ref": "#/definitions/GeoLocation"
        },
        "name": {
          "type": "string"
        },
        "ownerId": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "salads": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "sallyId": {
          "type": "string"
        }
      }
    },
...
    "Ingredient": {
      "type": "object",
      "properties": {
        "available": {
          "type": "boolean"
        },
        "category": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        },
        "ingredientId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "GeoLocation": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "number",
          "format": "double"
        },
        "longitude": {
          "type": "number",
          "format": "double"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "Salad": {
      "type": "object",
      "properties": {
        "available": {
          "type": "boolean"
        },
        "createdBy": {
          "type": "string"
        },
        "customSalad": {
          "type": "boolean"
        },
        "imageUrl": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "portions": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "saladId": {
          "type": "string"
        }
      }
    }
  }
}