# src/main/java/com/casabots/pantry/model/Order.java
/*******************************************************************************
 * Copyright (c) CASABOTS 2016
 * All Rights Reserved
 ******************************************************************************/

package com.casabots.pantry.model;

import org.springframework.data.annotation.Id;
import org.springframework.data.annotation.Transient;
import org.springframework.data.mongodb.core.mapping.Document;

import java.util.Date;

@Document(collection = "orders")
public class Order {

    @Id
    private String orderId;
    private String sallyId;
    private String userId;
    private String saladId;
    @Transient
    private boolean isOrderPlaced;
    private Date placedTime;

    public Order() {
    }

    public String getOrderId() {
        return orderId;
    }

    public void setOrderId(String orderId) {
        this.orderId = orderId;
    }

    public void setCustomSalad(boolean customSalad) {
    }

    public String getSallyId() {

        return sallyId;
    }

    public void setSallyId(String sallyId) {
        this.sallyId = sallyId;
    }

    public String getUserId() {
        return userId;
    }

    public void setUserId(String userId) {
        this.userId = userId;
    }

    public boolean isOrderPlaced() {
        return isOrderPlaced;
    }

    public void setOrderPlaced(boolean orderPlaced) {
        isOrderPlaced = orderPlaced;
    }

    public String getSaladId() {
        return saladId;
    }

    public void setSaladId(String saladId) {
        this.saladId = saladId;
    }

    public Date getPlacedTime() {
        return placedTime;
    }

    public void setPlacedTime(Date placedTime) {
        this.placedTime = placedTime;
    }
}